--- src/pages/signin.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import ScrollToTop from '../components/ScrollToTop';
-import SignIn from '../components/SignIn';
-
-const SigninPage = () => {
-    return (
-        <>
-            <ScrollToTop />
-            <SignIn />
-        </>
-    )
-}
-
-export default SigninPage;
