+++ src/pages/signin.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import ScrollToTop from '../components/ScrollToTop';
+import SignIn from '../components/SignIn';
+
+const SigninPage = () => {
+    return (
+        <>
+            <ScrollToTop />
+            <SignIn />
+        </>
+    )
+}
+
+export default SigninPage;
